refactor(macos): split pane tree and content views
